+++ Modules/Project.Windows.ProgramInfo/Public/Test-Environment.ps1
@@ -28,15 +28,29 @@ SOFTWARE.
 
 <#
 .SYNOPSIS
-Test if path is valid for firewall rule
+Test if a path is valid with additional checks
 
 .DESCRIPTION
-Same as Test-Path but expands environment variables, and checks if path is compatible
-for firewall rules
+Similar to Test-Path but expands environment variables and performs additional checks if desired:
+1. check if input path is compatible for firewall rules.
+2. check if the path leads to user profile
+Both of which can be limited to either container of leaf path type.
 
-.PARAMETER FilePath
+.PARAMETER Path
 Path to folder, Allows null or empty since input may come from other commandlets which
 can return empty or null
+
+.PARAMETER PathType
+A type of path to test, can be one of the following:
+1. Leaf -The path is file or registry entry
+2. Container - the path is container such as folder or registry key
+3. Any - Leaf or Container
+
+.PARAMETER Firewall
+Ensures the path is valid for firewall rule
+
+.PARAMETER UserProfile
+Checks if the path leads to user profile
 
 .EXAMPLE
 PS> Test-Environment %SystemDrive%
@@ -48,51 +62,90 @@ None. You cannot pipe objects to Test-Environment
 [bool] true if path exists, false otherwise
 
 .NOTES
-TODO: This should proably be inside utility module
+TODO: This should proably be part of utility module
 #>
 function Test-Environment
 {
-	[CmdletBinding(
+	[CmdletBinding(PositionalBinding = $false, DefaultParameterSetName = "None",
 		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Project.Windows.ProgramInfo/Help/en-US/Test-Environment.md")]
 	[OutputType([bool])]
 	param (
+		[Parameter(Position = 0, Mandatory = $true)]
+		[AllowNull()]
+		[AllowEmptyString()]
+		[string] $Path,
+
 		[Parameter()]
-		[string] $FilePath = $null
+		[ValidateSet("Leaf", "Container", "Any")]
+		[string] $PathType = "Container",
+
+		[Parameter()]
+		[switch] $Firewall,
+
+		[Parameter()]
+		[switch] $UserProfile
 	)
 
 	Write-Debug -Message "[$($MyInvocation.InvocationName)] params($($PSBoundParameters.Values))"
 	Write-Verbose -Message "[$($MyInvocation.InvocationName)] Checking if path is valid for firewall rule"
 
-	if ([string]::IsNullOrEmpty($FilePath))
+	if ([string]::IsNullOrEmpty($Path))
 	{
-		Write-Verbose -Message "[$($MyInvocation.InvocationName)] Returning false, file path is null or empty"
+		Write-Warning -Message "Path name is null or empty"
+		# Write-Verbose -Message "[$($MyInvocation.InvocationName)] Returning false, file path is null or empty"
 		return $false
 	}
+	elseif (Test-Path -Path $Path -IsValid)
+	{
+		$UserVariables = Get-EnvironmentVariable UserProfile | Select-Object -ExpandProperty Name
 
-	# TODO: We need target computer system drive instead of localmachine systemdrive
-	# NOTE: following would be mistake to use because paths without environment variables are fine
-	# ($FilePath -match "^$env:SystemDrive\\+Users(\\*$|\\+(?=\w+))") -or
-	if ([array]::Find($UserProfileEnvironment, [System.Predicate[string]] { $FilePath -like "$($args[0])*" }))
-	{
-		Write-Warning -Message "Path to program with environment variable that leads to user profile is not valid path for firewall"
+		# Status to check Path is valid with UserProfile and Firewall
+		[bool] $Status = $true
+
+		if ($UserProfile -or $Firewall)
+		{
+			if ([array]::Find($UserVariables, [System.Predicate[string]] { $Path -like "$($args[0])*" }))
+			{
+				if ($Firewall)
+				{
+					$Status = $false
+					Write-Warning -Message "Paths including environment variables which lead to user profile are not valid for firewall"
+				}
+			}
+			# TODO: We need target computer system drive instead of localmachine systemdrive
+			# NOTE: This pattern is fine for firewall
+			elseif ($UserProfile -and ($Path -notmatch "^$env:SystemDrive\\+Users(\\?$|\\+(?!Public\\.))"))
+			{
+				# NOTE: Not showing anything
+				Write-Verbose -Message "[$($MyInvocation.InvocationName)] The path does not lead to user profile"
+				return $false
+			}
+		}
+
+		if ($Status)
+		{
+			if (Test-Path -Path ([System.Environment]::ExpandEnvironmentVariables($Path)) -PathType $PathType)
+			{
+				return $true
+			}
+
+			# Why it failed:
+			$BadVariables = Get-EnvironmentVariable BlackList | Select-Object -ExpandProperty Name
+			if ([array]::Find($BadVariables, [System.Predicate[string]] { $Path -like "$($args[0])*" }))
+			{
+				Write-Warning -Message "Specified environment variable is not valid for paths"
+			}
+			else
+			{
+				Write-Warning -Message "Specified path does not exist"
+			}
+		}
 	}
-	elseif ([array]::Find($BlackListEnvironment, [System.Predicate[string]] { $FilePath -like "$($args[0])*" }))
+	else # -IsValid
 	{
-		Write-Warning -Message "Path to program with the specified environment variable is not valid for firewall"
-	}
-	elseif (!(Test-Path -Path $FilePath -IsValid))
-	{
-		Write-Warning -Message "Specified path is of invalid syntax"
-	}
-	elseif (!(Test-Path -Path ([System.Environment]::ExpandEnvironmentVariables($FilePath)) -PathType Container))
-	{
-		Write-Warning -Message "Specified path is not valid path to directory"
-	}
-	else
-	{
-		return $true
+		Write-Warning -Message "The path syntax is invalid"
 	}
 
-	Write-Information -Tags "Project" -MessageData "INFO: Invalid path is: $FilePath"
+	Write-Information -Tags "Project" -MessageData "INFO: Invalid path is: $Path"
 	return $false
 }